+++ CHANGELOG.md
@@ -7,6 +7,7 @@ project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 Feature:
 - Allow `output` to accept procs as defaults. (#188)
+- Allow any object that responds to `===` to be used as a type. (#187)
 - Allow objects without `Kernel` inclusion to function with the `play` DSL. (#180)
 
 ## v3.9.4